Add button to install adb (Android Debug Bridge) Can be found on: C:\ProgramData\chocolatey\lib\adb\tools\platform-tools - Fix "Enabling Windows Store apps Automatic Updates" error - Put ( ) after function names (my preference)

+++ src/scripts/optimize-privacy-and-performance.ps1
@@ -375,7 +375,9 @@ function Optimize-PrivacyAndPerformance() {
     If (!(Test-Path "$PathToLMPoliciesWindowsStore")) {
         New-Item -Path "$PathToLMPoliciesWindowsStore" -Force | Out-Null
     }
-    Remove-ItemProperty -Path "$PathToLMPoliciesWindowsStore" -Name "AutoDownload"
+    If ((Get-Item "$PathToLMPoliciesWindowsStore").GetValueNames() -like "AutoDownload") {
+        Remove-ItemProperty -Path "$PathToLMPoliciesWindowsStore" -Name "AutoDownload"
+    }
 
     Write-Section -Text "Power Plan Tweaks"
